+++ server/src/test/kotlin/com/adobe/testing/s3mock/controller/BucketControllerTest.kt
@@ -49,6 +49,7 @@ import com.adobe.testing.s3mock.dto.VersioningConfiguration
 import com.adobe.testing.s3mock.service.BucketService
 import com.adobe.testing.s3mock.service.MultipartService
 import com.adobe.testing.s3mock.service.ObjectService
+import com.adobe.testing.s3mock.service.VectorService
 import com.adobe.testing.s3mock.store.BucketMetadata
 import com.adobe.testing.s3mock.store.KmsKeyStore
 import com.adobe.testing.s3mock.util.AwsHttpHeaders.X_AMZ_BUCKET_LOCATION_NAME
@@ -80,7 +81,14 @@ import java.nio.file.Path
 import java.time.Instant
 
 @MockitoBean(
-  types = [KmsKeyStore::class, ObjectService::class, MultipartService::class, ObjectController::class, MultipartController::class],
+  types = [
+    KmsKeyStore::class,
+    ObjectService::class,
+    MultipartService::class,
+    VectorService::class,
+    ObjectController::class,
+    MultipartController::class,
+  ],
 )
 @WebMvcTest(
   controllers = [BucketController::class],