+++ server/src/test/kotlin/com/adobe/testing/s3mock/controller/FaviconControllerTest.kt
@@ -15,6 +15,7 @@
  */
 package com.adobe.testing.s3mock.controller
 
+import com.adobe.testing.s3mock.service.VectorService
 import com.adobe.testing.s3mock.store.KmsKeyStore
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
@@ -25,7 +26,9 @@ import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 
-@MockitoBean(types = [KmsKeyStore::class, ObjectController::class, BucketController::class, MultipartController::class])
+@MockitoBean(
+  types = [KmsKeyStore::class, VectorService::class, ObjectController::class, BucketController::class, MultipartController::class],
+)
 @WebMvcTest(controllers = [FaviconController::class])
 internal class FaviconControllerTest : BaseControllerTest() {
   @Autowired